동적 할당 가로/세로 변경

+++ PCA/PCA/PCA.h
@@ -29,7 +29,7 @@ void FileRead(uchar** img_in, int width, int height, char* path);
 void BMPtoRAW_Gray(uchar* in_bmp, uchar *out_raw, int width, int height);
 
 void GetAverageMatrix(uchar** img_ori, uchar** ave_ori, int width, int height);
-void GetCovarianceMat(uchar** img_ori, uchar** img_ave, uchar** img_cov, int width, int height);
+void GetCovarianceMat(uchar** img_ori, uchar** img_ave, int** img_cov, int width, int height);
 void Transpose(int** img_ave_removed, int** img_ave_removed_T, int width, int height);
